Add snake.core.tick
This will be our main game loop. At the moment it combines a new "direction"
property in the game state with a the current state of the snake, and updates
the state with the next snake.

As part of this change, never call (main) it from the CLJS - put it in the HTML.
This is because we don't want a loop to start on each reload.