Refactor and improve collision detection
There was a bug in the snake-colliding? method: the snake is a vector
with the head at the END, not the beginning. Therefore it only worked
when the end of the snake crossed over the rest of the body.

While fixing this, I created a next-state method which seems to improve
tick significantly. Maps returned from next-state can declare the state
changes they want and it's trivial to apply them using merge, rather
than have if-statements which make 1..n calls to assoc.