Add speed controls
Change the :speed value in the main state map to be an integer which we
map to a millisecond value, used as the delay before the next tick.

At first I thought the input element had to live in the 'core' namespace
as it needs to hand its command to 'handle-command', but it seems
passing a function is a nice way to export the necessary bits of scope
from core.